Refactor tv controller to fix chat mixin logic
