+++ lib/src/model/tv/tv_controller.dart
@@ -1,28 +1,12 @@
 import 'dart:async';
 
 import 'package:dartchess/dartchess.dart';
-import 'package:deep_pick/deep_pick.dart';
-import 'package:fast_immutable_collections/fast_immutable_collections.dart';
-import 'package:flutter/foundation.dart' show VoidCallback;
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:lichess_mobile/src/model/chat/chat_mixin.dart';
-import 'package:lichess_mobile/src/model/common/chess.dart';
 import 'package:lichess_mobile/src/model/common/id.dart';
-import 'package:lichess_mobile/src/model/common/service/sound_service.dart';
-import 'package:lichess_mobile/src/model/common/socket.dart';
-import 'package:lichess_mobile/src/model/game/game.dart';
-import 'package:lichess_mobile/src/model/game/game_socket_events.dart';
-import 'package:lichess_mobile/src/model/game/game_status.dart';
-import 'package:lichess_mobile/src/model/game/material_diff.dart';
-import 'package:lichess_mobile/src/model/game/playable_game.dart';
 import 'package:lichess_mobile/src/model/tv/tv_channel.dart';
+import 'package:lichess_mobile/src/model/tv/tv_game_controller.dart';
 import 'package:lichess_mobile/src/model/tv/tv_repository.dart';
-import 'package:lichess_mobile/src/model/tv/tv_socket_events.dart';
 import 'package:lichess_mobile/src/model/user/user_repository.dart';
-import 'package:lichess_mobile/src/network/socket.dart';
-
-part 'tv_controller.freezed.dart';
 
 typedef TvControllerParams = ({
   TvChannel? channel,
@@ -31,63 +15,44 @@ typedef TvControllerParams = ({
 });
 
 final tvControllerProvider = AsyncNotifierProvider.autoDispose
-    .family<TvController, TvState, TvControllerParams>(
+    .family<TvController, TvGameControllerParams, TvControllerParams>(
       TvController.new,
       name: 'TvControllerProvider',
     );
 
-class TvController extends AsyncNotifier<TvState> with ChatMixin<TvState> {
+/// Resolves which game to watch on the TV screen and moves to the next game
+/// when the watched channel switches.
+///
+/// The actual game (board, moves, chat) is handled by the [TvGameController]
+/// identified by the [TvGameControllerParams] exposed as this controller's
+/// state.
+class TvController extends AsyncNotifier<TvGameControllerParams> {
   TvController(this.params);
 
   final TvControllerParams params;
 
-  StreamSubscription<SocketEvent>? _socketSubscription;
-
-  VoidCallback? _onReload;
-
   @override
-  @protected
-  StringId get chatId => state.requireValue.game.id;
-
-  @override
-  @protected
-  String get chatReportResource => 'game/$chatId';
-
-  @override
-  @protected
-  bool get chatIsPublic => true;
-
-  @override
-  Future<TvState> build() {
+  Future<TvGameControllerParams> build() {
     assert(
-      params.channel != null || params.userId != null,
-      'Either a channel or a userId must be provided',
+      params.channel != null || params.userId != null || params.initialGame != null,
+      'Either a channel, a userId or an initial game must be provided',
     );
 
-    _onReload = ref.invalidateSelf;
-
-    ref.onDispose(() {
-      _socketSubscription?.cancel();
-      _onReload = null;
-    });
-
-    return _connectWebsocket(params.initialGame);
+    return _resolveGame(params.initialGame);
   }
 
-  SoundService get _soundService => ref.read(soundServiceProvider);
+  /// Called when the watched channel selects a new game.
+  void moveToNextGame((GameId id, Side orientation) game) {
+    state = AsyncValue.data((gameId: game.$1, orientation: game.$2, source: params));
+  }
 
-  @override
+  /// Re-resolves the current game (the channel may have switched while the app
+  /// was in the background).
   Future<void> onFocusRegained() async {
-    final newState = await _connectWebsocket(null);
-    state = AsyncValue.data(newState);
+    state = AsyncValue.data(await _resolveGame(null));
   }
 
-  @override
-  void onForegroundLost() {
-    _socketSubscription?.cancel();
-  }
-
-  Future<TvState> _connectWebsocket((GameId id, Side orientation)? game) async {
+  Future<TvGameControllerParams> _resolveGame((GameId id, Side orientation)? game) async {
     GameId id;
     Side orientation;
 
@@ -100,246 +65,14 @@ class TvController extends AsyncNotifier<TvState> with ChatMixin<TvState> {
       id = channelGame.id;
       orientation = channelGame.side ?? Side.white;
     } else if (params.userId != null) {
-      final game = await ref.read(userRepositoryProvider).getCurrentGame(params.userId!);
-      id = game.id;
-      orientation = game.playerSideOf(params.userId!) ?? Side.white;
+      final userGame = await ref.read(userRepositoryProvider).getCurrentGame(params.userId!);
+      id = userGame.id;
+      orientation = userGame.playerSideOf(params.userId!) ?? Side.white;
     } else {
-      id = state.value?.game.id ?? params.initialGame!.$1;
+      id = state.value?.gameId ?? params.initialGame!.$1;
       orientation = state.value?.orientation ?? params.initialGame!.$2;
     }
 
-    final socketClient = ref
-        .read(socketPoolProvider)
-        .open(
-          Uri(
-            path: '/watch/$id/${orientation.name}/v6',
-            queryParameters: params.userId != null ? {'userTv': params.userId.toString()} : null,
-          ),
-          forceReconnect: true,
-          onEventGapFailure: () {
-            _onReload?.call();
-          },
-        );
-
-    _socketSubscription?.cancel();
-    _socketSubscription = socketClient.stream.listen(handleSocketEvent);
-
-    return socketClient.stream.firstWhere((e) => e.topic == 'full').then((event) {
-      final fullEvent = GameFullEvent.fromJson(event.data as Map<String, dynamic>);
-      socketClient.version = fullEvent.socketEventVersion;
-
-      Future(() async {
-        if (state.hasValue) {
-          state = AsyncValue.data(
-            state.requireValue.copyWith(chatState: await initChat(state.requireValue.game.chat)),
-          );
-        }
-      });
-
-      return TvState(
-        game: fullEvent.game,
-        stepCursor: fullEvent.game.steps.length - 1,
-        orientation: orientation,
-      );
-    });
+    return (gameId: id, orientation: orientation, source: params);
   }
-
-  Future<void> _moveToNextGame((GameId id, Side orientation) game) async {
-    final newState = await _connectWebsocket(game);
-    state = AsyncValue.data(newState);
-  }
-
-  bool canGoBack() => state.mapOrNull(data: (d) => d.value.stepCursor > 0) ?? false;
-
-  bool canGoForward() =>
-      state.mapOrNull(data: (d) => d.value.stepCursor < d.value.game.steps.length - 1) ?? false;
-
-  void toggleBoard() {
-    if (state.hasValue) {
-      final curState = state.requireValue;
-      state = AsyncValue.data(curState.copyWith(orientation: curState.orientation.opposite));
-    }
-  }
-
-  void cursorForward() {
-    if (state.hasValue) {
-      final curState = state.requireValue;
-      if (curState.stepCursor < curState.game.steps.length - 1) {
-        state = AsyncValue.data(curState.copyWith(stepCursor: curState.stepCursor + 1));
-        final san = curState.game.stepAt(curState.stepCursor + 1).sanMove?.san;
-        if (san != null) {
-          _playReplayMoveSound(san);
-        }
-      }
-    }
-  }
-
-  void cursorBackward() {
-    if (state.hasValue) {
-      final curState = state.requireValue;
-      if (curState.stepCursor > 0) {
-        state = AsyncValue.data(curState.copyWith(stepCursor: curState.stepCursor - 1));
-        final san = curState.game.stepAt(curState.stepCursor - 1).sanMove?.san;
-        if (san != null) {
-          _playReplayMoveSound(san);
-        }
-      }
-    }
-  }
-
-  void goToMove(int index) {
-    if (state.hasValue) {
-      final curState = state.requireValue;
-      if (index >= 0 && index < curState.game.steps.length) {
-        state = AsyncValue.data(curState.copyWith(stepCursor: index));
-      }
-    }
-  }
-
-  void _playReplayMoveSound(String san) {
-    final soundService = ref.read(soundServiceProvider);
-    if (san.contains('x')) {
-      soundService.play(Sound.capture);
-    } else {
-      soundService.play(Sound.move);
-    }
-  }
-
-  @protected
-  @override
-  void updateChatState(ChatState newState) {
-    state = AsyncValue.data(state.requireValue.copyWith(chatState: newState));
-  }
-
-  @protected
-  @override
-  void handleSocketEvent(SocketEvent event) {
-    super.handleSocketEvent(event);
-
-    if (!state.hasValue) {
-      return;
-    }
-
-    switch (event.topic) {
-      case 'resync':
-        _onReload?.call();
-      case 'reload':
-        if (event.data is Map<String, dynamic>) {
-          final data = event.data as Map<String, dynamic>;
-          if (data['t'] == null) {
-            _onReload?.call();
-            return;
-          }
-          final reloadEvent = SocketEvent(topic: data['t'] as String, data: data['d']);
-          handleSocketEvent(reloadEvent);
-        } else {
-          _onReload?.call();
-        }
-      case 'crowd':
-        final data = event.data as Map<String, dynamic>;
-        final watcherData = data['watchers'];
-        if (watcherData != null && watcherData is Map<String, dynamic>) {
-          final nb = watcherData['nb'] as int? ?? 0;
-          final users =
-              (watcherData['users'] as List<dynamic>?)?.map((e) => e.toString()).toIList() ??
-              const IList.empty();
-          state = AsyncData(state.requireValue.copyWith(nbWatchers: nb, watcherNames: users));
-        }
-      case 'rematchTaken':
-        _onReload?.call();
-      case 'move' || 'drop':
-        final curState = state.requireValue;
-        final data = MoveEvent.fromJson(event.data as Map<String, dynamic>);
-        final lastPos = curState.game.lastPosition;
-        final move = Move.parse(data.uci)!;
-        final sanMove = SanMove(data.san, move);
-        final newPos = lastPos.playUnchecked(move);
-        final newStep = GameStep(
-          sanMove: sanMove,
-          position: newPos,
-          diff: MaterialDiff.fromPosition(newPos),
-        );
-
-        TvState newState = curState.copyWith(
-          game: curState.game.copyWith(steps: curState.game.steps.add(newStep)),
-        );
-
-        if (newState.game.clock != null && data.clock != null) {
-          newState = newState.copyWith.game.clock!(
-            white: data.clock!.white,
-            black: data.clock!.black,
-            lag: data.clock!.lag,
-            at: data.clock!.at,
-          );
-        }
-        if (!curState.isReplaying) {
-          newState = newState.copyWith(stepCursor: newState.stepCursor + 1);
-
-          if (data.san.contains('x')) {
-            _soundService.play(Sound.capture);
-          } else {
-            _soundService.play(Sound.move);
-          }
-        }
-
-        state = AsyncData(newState);
-
-      case 'endData':
-        final endData = GameEndEvent.fromJson(event.data as Map<String, dynamic>);
-        TvState newState = state.requireValue.copyWith(
-          game: state.requireValue.game.copyWith(status: endData.status, winner: endData.winner),
-        );
-        if (endData.clock != null) {
-          newState = newState.copyWith.game.clock!(
-            white: endData.clock!.white,
-            black: endData.clock!.black,
-            at: DateTime.now(),
-            lag: null,
-          );
-        }
-        state = AsyncData(newState);
-
-      case 'tvSelect':
-        final json = event.data as Map<String, dynamic>;
-        final eventChannel = pick(json, 'channel').asTvChannelOrNull();
-        if (eventChannel != null && eventChannel == params.channel) {
-          final data = TvSelectEvent.fromJson(json);
-          _moveToNextGame((data.id, data.orientation));
-        }
-    }
-  }
-}
-
-@freezed
-sealed class TvState with _$TvState, ChatMixinState {
-  const TvState._();
-
-  const factory TvState({
-    required PlayableGame game,
-    required int stepCursor,
-    required Side orientation,
-    ChatState? chatState,
-    @Default(0) int nbWatchers,
-    @Default(IList<String>.empty()) IList<String> watcherNames,
-  }) = _TvState;
-
-  bool get isReplaying => stepCursor < game.steps.length - 1;
-
-  Side? get activeClockSide {
-    if (game.clock == null) {
-      return null;
-    }
-
-    if (game.status == GameStatus.started) {
-      final pos = game.lastPosition;
-      if (pos.fullmoves > 1) {
-        return pos.turn;
-      }
-    }
-
-    return null;
-  }
-
-  @override
-  bool get chatEnabled => true;
 }
